Add support for command choices and image options in `gen` command
- Add command choices for the `gen` command to specify the DALL-E model, image size, and quality.
- Modify the `gen` command to accept additional optional arguments for image options.
- Update the `Pal9000` class to pass the image options to the `PixelPainter` and `CostSpy` classes.
- Modify the `PixelPainter` class to use the image options when generating images.
- Modify the `CostSpy` class to accept the image options as a keyword argument in the `process` method.

This commit adds support for command choices and image options in the `gen` command, allowing users to customize the DALL-E model, image size, and quality. The `Pal9000` class is updated to pass the image options to the `PixelPainter` and `CostSpy` classes. The `PixelPainter` class is modified to use the image options when generating images. The `CostSpy` class now accepts the image options as a keyword argument in the `process` method.